Revert "Move transaction_types to wasm_bindings"
This reverts commit 02e1ece833d14ffeac74d110692cda8b9af0cdd7.
