+++ quadratic-core/src/wasm_bindings/mod.rs
@@ -7,8 +7,6 @@ pub mod js;
 pub mod lsp;
 pub mod pos;
 pub mod rect;
-pub mod sheet_offsets_wasm;
-pub mod transaction_types;
 
 use crate::controller::GridController;
 use crate::grid::*;
